% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_convert_logicals.R
\name{auto_convert_logicals}
\alias{auto_convert_logicals}
\title{Convert all psuedo-logical columns in a dataframe to booleans}
\usage{
auto_convert_logicals(input_data)
}
\arguments{
\item{input_data}{dataframe}
}
\value{
dataframe
}
\description{
Convert all psuedo-logical columns in a dataframe to booleans
}
\examples{
x <- data.frame(
   zero_one = sample(0:1, 10, replace = TRUE),
   TF = sample(c("T", "F"), 10, replace = TRUE),
   truefalse = sample(c('true', 'false'), 10, replace = TRUE)
 )
auto_convert_logicals(x)
}
\author{
Joe Marlo
}
