% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_data_types.R
\name{clean_data_types}
\alias{clean_data_types}
\title{Simplify data types for humans}
\usage{
clean_data_types(input_data)
}
\arguments{
\item{input_data}{dataframe}
}
\value{
character vector of length ncol(input_data)
}
\description{
Categorizes R data types into 'Continuous', 'Categorical', or 'Binary'
}
\examples{
n_row <- 10
my_character = sample(c('one', 'two', 'three'), size = n_row, replace = TRUE)
my_logical = sample(c(TRUE, FALSE), size = n_row, replace = TRUE)
my_numeric = rnorm(n_row)
X <- data.frame(my_character = my_character, my_logical = my_logical, my_numeric = my_numeric)
clean_data_types(X)
}
