% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_detect_plot_vars.R
\name{clean_detect_plot_vars}
\alias{clean_detect_plot_vars}
\title{Determine which default columns to use in plot_exploration()}
\usage{
clean_detect_plot_vars(.column_types, .treatment_column, .response_column)
}
\arguments{
\item{.column_types}{a list containing names of categorical and continuous columns. Ideally from output of clean_detect_column_types().}

\item{.treatment_column}{name of column denoting treatment}

\item{.response_column}{name of the outcome column}
}
\value{
a list denoting which columns to use in plotting
}
\description{
Determine which default columns to use in plot_exploration()
}
\examples{
X <- data.frame(
 treatment = sample(as.logical(0:1), 5, TRUE), 
 response = rnorm(5),
 X1 = 1:5, 
 X2 = rnorm(5), 
 X3 = LETTERS[1:5], 
 X4 = as.factor(LETTERS[1:5])
)
column_types <- clean_detect_column_types(X)
clean_detect_plot_vars(column_types, 'treatment', 'response')
}
\seealso{
\code{\link{plot_exploration} \link{clean_detect_column_types}}
}
