% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_detect_ZYX_columns.R
\name{clean_detect_ZYX_columns}
\alias{clean_detect_ZYX_columns}
\title{Attempt to detect which columns of a dataframe are Z, Y, and X}
\usage{
clean_detect_ZYX_columns(input_colnames)
}
\arguments{
\item{input_colnames}{character vector of column names}
}
\value{
a list denoting which column names are Z, Y, and X
}
\description{
Attempts to auto detect which columns are treatment and response based on the column name alone. If none are detected, then returns the first two respectivel
}
\examples{
x <- c("treatment", "rsp", "dummy1", "dummy2")
clean_detect_ZYX_columns(x)
}
\author{
Joe Marlo
}
