% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_type_to_complex.R
\name{convert_data_type_to_complex}
\alias{convert_data_type_to_complex}
\title{Convert human readable data type to R data}
\usage{
convert_data_type_to_complex(.data, .simple_data_types)
}
\arguments{
\item{.data}{a dataframe}

\item{.simple_data_types}{human readable data types from convert_data_type_to_simple. Must be length == ncol(.data)}
}
\value{
a dataframe with dim == dim(.data)
}
\description{
The inverse of convert_data_type_to_simple(). Converts a dataframe to new data types based on human readable strings.
}
\seealso{
\code{\link{convert_data_type_to_simple}}
}
\author{
Joe Marlo
}
