% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datatable.R
\name{create_datatable}
\alias{create_datatable}
\title{Create a pretty datatable}
\usage{
create_datatable(...)
}
\arguments{
\item{...}{arguments to be passed to DT::datatable. Typically a dataframe}
}
\value{
HTML datatable
}
\description{
Create an HTML datatable using the JavaScript library DataTables with predefined defaults
}
\examples{
X <- data.frame(x = 1:10, y = 1:10)
create_datatable(X)
}
\author{
Joe Marlo
}
