% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ZYX_columns.R
\name{detect_ZYX_columns}
\alias{detect_ZYX_columns}
\title{Attempt to detect which columns of a dataframe are Z, Y, and X}
\usage{
detect_ZYX_columns(input_colnames)
}
\arguments{
\item{input_colnames}{character vector of column names}
}
\value{
a list denoting which column names are Z, Y, and X
}
\description{
Attempt to detect which columns of a dataframe are Z, Y, and X
}
\examples{
x <- c("treatment", "rsp", "dummy1", "dummy2")
detect_ZYX_columns(x)
}
\author{
Joe Marlo
}
