% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot the balance}
\usage{
plot_balance(.data, selected_cols)
}
\arguments{
\item{selected_cols}{character list of column names to plot. Typically input$analysis_plot_balance_select_var}

\item{X}{dataframe formatted with Z, Y, and X variables. Typically store$selected_df}
}
\value{
ggplot object
}
\description{
Returns a ggplot balance plot
}
\author{
Joe Marlo
}
