% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_common_support_sd.R
\name{plot_common_support_sd}
\alias{plot_common_support_sd}
\title{Plot common support based on the standard deviation rule}
\usage{
plot_common_support_sd(.model)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}
}
\value{
ggplot object
}
\description{
Returns a ggplot common support plot
}
\author{
George Perrett, Joe Marlo
}
