% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DAG.R
\name{plot_DAG}
\alias{plot_DAG}
\title{Plot a simple DAG}
\usage{
plot_DAG(Z_col, Y_col, X_cols)
}
\arguments{
\item{Z_col}{string}

\item{Y_col}{string}

\item{X_cols}{string or list of strings}
}
\value{
ggplot object
}
\description{
Plot a simple DAG
}
\examples{
plot_DAG("myZ", "myY", c("X1", "X2", "X3"))
}
