% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic_common_support.R
\name{plot_diagnostic_common_support}
\alias{plot_diagnostic_common_support}
\title{Plot common support based on the standard deviation rule, chi squared rule or both}
\usage{
plot_diagnostic_common_support(.model, .rule = c("none", "sd", "chi"))
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}
}
\value{
ggplot object
}
\description{
Returns a ggplot common support plot
}
\author{
George Perrett, Joe Marlo
}
