% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploration.R
\name{plot_exploration}
\alias{plot_exploration}
\title{General purpose plotting for EDA}
\usage{
plot_exploration(
  .data,
  .plot_type = c("Pairs", "Scatter", "Histogram", "Density", "Boxplot"),
  .x,
  .y,
  .fill,
  .fill_static = "#5c5980",
  .size,
  .alpha = 0.9,
  .vars_pairs,
  .n_bins,
  .jitter,
  .groups,
  .facet,
  .facet_second,
  .include_regression = c("Include", "None")
)
}
\arguments{
\item{.data}{typically store$selected_df}

\item{.plot_type}{one of c("Pairs", 'Scatter', 'Histogram', 'Density', 'Boxplot')}

\item{.x}{}

\item{.y}{}

\item{.fill}{}

\item{.fill_static}{}

\item{.size}{}

\item{.alpha}{}

\item{.vars_pairs}{}

\item{.n_bins}{}

\item{.jitter}{}

\item{.groups}{}

\item{.facet}{}

\item{.facet_second}{}

\item{.include_regression}{one of c("Include", "None")}
}
\value{

}
\description{
To be used within Shiny
}
