% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ITE.R
\name{plot_ITE}
\alias{plot_ITE}
\title{Plot histogram of individual treatment effects}
\usage{
plot_ITE(.model)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}
}
\value{
ggplot object
}
\description{
Returns a ggplot ITE plot
}
\author{
George Perrett, Joe Marlo
}
