% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_estimate.R
\name{plot_model_estimate}
\alias{plot_model_estimate}
\title{Plot the estimands of a bartCause::bartc() model}
\usage{
plot_model_estimate(model)
}
\arguments{
\item{model}{a bartCause::bartc() model, typically store$model_results}
}
\value{
ggplot object
}
\description{
Returns a ggplot of the estimated effect
}
\author{
Joe Marlo, George Perrett
}
