% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap_pScores.R
\name{plot_overlap_pScores}
\alias{plot_overlap_pScores}
\title{Plot the overlap via propensity score method}
\usage{
plot_overlap_pScores(.data, plt_type)
}
\arguments{
\item{.data}{dataframe. Typically store$selected_df}
}
\value{
ggplot object
}
\description{
Plot histograms showing the overlap between propensity scores by treatment status.
}
\author{
George Perrett
}
