% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Trace plot the estimands of a bartCause::bartc() model}
\usage{
plot_trace(.model)
}
\arguments{
\item{.model}{a bartCause::bartc() model, typically store$model_results}
}
\value{
ggplot object
}
\description{
Returns a ggplot of the stimated effect (i.e. trace plot).
}
\author{
Joe Marlo, George Perrett
}
