% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot}
\usage{
save_plot(plot, filename, height = 8, width = 8)
}
\arguments{
\item{plot}{A ggplot object.}

\item{filename}{Output file path, including file extension.}

\item{height}{Plot height in inches.}

\item{width}{Plot width in inches.}
}
\value{
Invisibly returns the input filename.
}
\description{
Saves ggplot objects to disk.
}
