% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datatable.R
\name{create_datatable}
\alias{create_datatable}
\title{Create a pretty datatable}
\usage{
create_datatable(...)
}
\arguments{
\item{...}{arguments to be passed to DT::datatable. Typically a dataframe}
}
\value{
HTML datatable
}
\description{
Create an HTML datatable using the JavaScript library DataTables. This is a wrapper around DT::datatable with commonly used arguments set as defaults.
}
\examples{
X <- data.frame(x = 1:10, y = 1:10)
create_datatable(X)

## within Shiny
# UI
#DT::dataTableOutput(outputId = tableId)

# server
#output$tableId <- DT::renderDataTable(create_datatable(X))
}
\author{
Joe Marlo
}
