% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log.R
\name{create_log}
\alias{create_log}
\title{Build a reproducible 'log' which mimics the thinkCausal work-flow}
\usage{
create_log()
}
\value{
string
}
\description{
Build a reproducible 'log' which mimics the thinkCausal work-flow
}
