% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_progress_bar.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create a progress bar}
\usage{
create_progress_bar(progress)
}
\arguments{
\item{progress}{numeric between 0 an 1}
}
\value{
HTML
}
\description{
Create a bootstrap HTML progress bar filled to `progress` percentage
}
\examples{
create_progress_bar(0.8)
}
\author{
Joe Marlo
}
