% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cate_test.R
\name{plot_cate_test}
\alias{plot_cate_test}
\title{Variable importance of Bayesian Additive Regression Trees}
\usage{
plot_cate_test(.model, confounders)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}

\item{confounders}{matrix of confounders}
}
\value{
a list containing variable importance plot & plots for each conditional effect
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & conditional effects plots.
}
\author{
George Perrett, Joe Marlo
}
