% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap_vars.R
\name{plot_overlap_vars}
\alias{plot_overlap_vars}
\title{Plot the overlap of variables}
\usage{
plot_overlap_vars(.data, selected_cols)
}
\arguments{
\item{.data}{dataframe. Typically store$selected_df}

\item{selected_cols}{list of columns from .data to plot}
}
\value{
ggplot object
}
\description{
Plot histograms showing the overlap between variables by treatment status.
}
\author{
George Perrett, Joe Marlo
}
